feat: 게시글 좋아요 권한 설정 및 프론트 일부 오류 개선

+++ backend/rush/src/main/java/rush/rush/service/FindArticleService.java
@@ -61,7 +61,7 @@ public class FindArticleService {
         User author = article.getUser();
         AuthorResponse authorResponse = new AuthorResponse(author.getId(),
             author.getNickName(), author.getImageUrl());
-        int totalLikes = articleLikeRepository.countByArticleId(article.getId());
+        Long totalLikes = articleLikeRepository.countByArticleId(article.getId());
 
         return new ArticleResponse(
             article.getId(),
